refactor: extract the input layer into src/input/
main.rs 2759 -> 2195. With the render tree in src/ui/ and the HTTP layer in
src/api/, the event handlers were the last big block standing between the app
loop and the state it mutates.

  key.rs     the main keymap, and drives_library
  mouse.rs   clicks, drags and wheel, resolved against FrameOut hit rects
  actions.rs input while the actions overlay is open, plus copy-to-clipboard
  media.rs   OS media keys (souvlaki)

This is the layer that runs the other way from ui/: ui reads &App and never
writes, input writes App and never draws. It reads FrameOut only for the hit
rects the renderer left behind. Stating that in both module docs makes the
loop legible in a way it wasn't when all three lived in one file.

Pure move. No logic changed - only pub(crate) visibility, use statements, and
rustfmt rewrapping the handle_key and consume_media_event signatures that
pub(crate) pushed past 100 columns. Every in-body comment moved with its body,
including the seek state machine, the hit-rect guards, and the media-control
open flag.

drives_library and consume_media_event stay reachable from main_tests/ through
the re-export glob. 139 tests still pass; fmt, clippy and test clean with
--all-features under -D warnings.
