+++ src/main.rs
@@ -9,6 +9,10 @@
 /// Lives in the binary (not the library) because it speaks the model types
 /// defined here.
 mod api;
+/// The input layer. Turns terminal and media-key events into `App` mutations
+/// and channel sends — the one layer that writes state.
+/// Lives in the binary (not the library) because it mutates `App`, which is here.
+mod input;
 /// The render tree. Reads `App`, writes `FrameOut`; never the other way round.
 /// Lives in the binary (not the library) because it needs `App`, which is here.
 mod ui;
@@ -32,6 +36,7 @@ use ratatui::Frame;
 use ratatui_image::picker::Picker;
 
 use api::*;
+use input::*;
 use myx::anim::ThemeFade;
 use myx::audio::NUM_BANDS;
 use myx::components::{gradient_line, gradient_progress, left_bar_block};
@@ -1598,575 +1603,6 @@ fn play_selected_context(app: &mut App, shuffle: bool) {
         Some((uri, name)) => app.play_context_row(uri, name, shuffle),
         None => app.status = "not a playlist, album, or artist".to_string(),
     }
-}
-
-/// Mouse input. Mirrors `handle_key`: returns `true` when the app should quit.
-fn handle_mouse(
-    app: &mut App,
-    out: &FrameOut,
-    m: crossterm::event::MouseEvent,
-    chans: &UiChannels,
-) -> bool {
-    if matches!(
-        m.kind,
-        MouseEventKind::Down(MouseButton::Left) | MouseEventKind::Drag(MouseButton::Left)
-    ) {
-        let is_down = matches!(m.kind, MouseEventKind::Down(MouseButton::Left));
-        let mut consumed = false;
-        // Drag the sidebar scrollbar (2-col grab target) to scroll.
-        if let Some(sb) = out.hits.scroll {
-            if m.column + 1 >= sb.x
-                && m.column <= sb.x
-                && m.row >= sb.y
-                && m.row < sb.y + sb.height
-                && sb.height > 0
-            {
-                consumed = true;
-                let total = out.hits.scroll_len;
-                if total > 1 {
-                    let denom = sb.height.saturating_sub(1).max(1) as f32;
-                    let frac = (m.row - sb.y) as f32 / denom;
-                    let sel = (frac * (total - 1) as f32).round() as usize;
-                    app.browse.selected = sel.min(total - 1);
-                    app.normalize_selection();
-                }
-            }
-        }
-        // Click/drag the volume meter bars to set volume.
-        if !consumed {
-            if let Some(vr) = out.hits.vol {
-                if m.row == vr.y && m.column >= vr.x && m.column < vr.x + vr.width && vr.width > 0 {
-                    consumed = true;
-                    let offset = (m.column - vr.x) as u32;
-                    let vol = (((offset + 1) * 100) / vr.width as u32).min(100) as u8;
-                    app.transport.volume = vol;
-                    let _ = app.svc.engine.set_volume(vol_u16(app.transport.volume));
-                }
-            }
-        }
-        // Otherwise an initial click on the progress bar seeks.
-        if !consumed && is_down {
-            if let Some(bar) = out.hits.bar {
-                if m.row == bar.y
-                    && m.column >= bar.x
-                    && m.column < bar.x + bar.width
-                    && bar.width > 0
-                {
-                    if let Some(dur) = app.playback.now.as_ref().map(|n| n.duration_ms) {
-                        let frac = (m.column - bar.x) as f32 / bar.width as f32;
-                        app.playback
-                            .seek_to(&app.svc.engine, (frac * dur as f32) as u32);
-                    }
-                }
-            }
-        }
-        // View-tab click -> switch the right pane.
-        if !consumed && is_down {
-            let hit = out
-                .hits
-                .tabs
-                .iter()
-                .find(|(_, r)| m.row == r.y && m.column >= r.x && m.column < r.x + r.width)
-                .map(|(v, _)| *v);
-            if let Some(v) = hit {
-                app.view.mode = v;
-                consumed = true;
-            }
-        }
-        // Library click -> select; double-click (same row <400ms) -> activate.
-        if !consumed && is_down {
-            if let Some(lr) = out.hits.lib {
-                if m.column >= lr.x
-                    && m.column < lr.x + lr.width
-                    && m.row >= lr.y
-                    && m.row < lr.y + lr.height
-                {
-                    let idx = out.lib_offset + (m.row - lr.y) as usize;
-                    let selectable = app
-                        .cur_items()
-                        .get(idx)
-                        .map(|it| !it.is_header)
-                        .unwrap_or(false);
-                    if selectable {
-                        app.browse.selected = idx;
-                        let now = Instant::now();
-                        let dbl = app
-                            .session
-                            .last_click
-                            .map(|(r0, t0)| {
-                                r0 == m.row && now.duration_since(t0) < Duration::from_millis(400)
-                            })
-                            .unwrap_or(false);
-                        if dbl {
-                            app.session.last_click = None;
-                            let quit =
-                                handle_key(app, KeyCode::Enter, KeyModifiers::empty(), chans);
-                            if quit {
-                                return true;
-                            }
-                        } else {
-                            app.session.last_click = Some((m.row, now));
-                        }
-                    }
-                }
-            }
-        }
-    }
-    // Scroll wheel → volume (anywhere in the window).
-    if matches!(
-        m.kind,
-        MouseEventKind::ScrollUp | MouseEventKind::ScrollDown
-    ) {
-        match m.kind {
-            MouseEventKind::ScrollUp => {
-                app.transport.volume = (app.transport.volume + 5).min(100);
-                let _ = app.svc.engine.set_volume(vol_u16(app.transport.volume));
-            }
-            MouseEventKind::ScrollDown => {
-                app.transport.volume = app.transport.volume.saturating_sub(5);
-                let _ = app.svc.engine.set_volume(vol_u16(app.transport.volume));
-            }
-            _ => {}
-        }
-    }
-    false
-}
-
-/// Returns true if the app should quit.
-fn handle_key(app: &mut App, code: KeyCode, mods: KeyModifiers, chans: &UiChannels) -> bool {
-    // --- Actions menu captures input while open ---
-    // Double-press Ctrl-C to quit (works from anywhere). Single press arms it.
-    if code == KeyCode::Char('c') && mods.contains(KeyModifiers::CONTROL) {
-        let now = Instant::now();
-        if app
-            .session
-            .last_ctrl_c
-            .map(|t| now.duration_since(t) < Duration::from_millis(1500))
-            .unwrap_or(false)
-        {
-            return true;
-        }
-        app.session.last_ctrl_c = Some(now);
-        app.status = "press Ctrl-C again to quit".to_string();
-        return false;
-    }
-
-    if app.view.actions.is_some() {
-        handle_action_key(app, code, &chans.detail, &chans.astatus);
-        return false;
-    }
-
-    // --- Search input mode captures everything ---
-    if app.search.input_mode {
-        match code {
-            KeyCode::Esc => app.search.input_mode = false,
-            KeyCode::Enter => {
-                app.search.input_mode = false;
-                let q = app.search.query.trim().to_string();
-                if !q.is_empty() {
-                    app.search.searching = true;
-                    app.browse.selected = 0;
-                    app.status = "searching…".to_string();
-                    spawn_search(app.svc.webapi.clone(), q, chans.search.clone());
-                }
-            }
-            KeyCode::Backspace => {
-                app.search.query.pop();
-            }
-            KeyCode::Char(c) => app.search.query.push(c),
-            _ => {}
-        }
-        return false;
-    }
-
-    // Zen hides the library, so the keys that drive one do nothing rather than
-    // moving a selection nobody can see. Placed after the overlays above, which
-    // stay usable if one was already open when zen came on.
-    if app.view.zen && drives_library(code) {
-        return false;
-    }
-
-    match code {
-        KeyCode::Char('/') => {
-            app.search.input_mode = true;
-            app.search.query.clear();
-        }
-        KeyCode::Char('q') => return true,
-        KeyCode::Esc => {
-            if let Some(d) = app.browse.details.pop() {
-                app.browse.selected = d.parent_selected;
-            } else if app.search.searching {
-                app.search.searching = false;
-                app.browse.selected = 0;
-            }
-            // Nothing to back out of — Esc no longer quits (use q or Ctrl-C twice).
-        }
-        KeyCode::Char(' ') | KeyCode::Char('p') | KeyCode::Media(MediaKeyCode::PlayPause) => {
-            if app.transport.playback_started {
-                let _ = app.svc.engine.toggle();
-            } else if app.session.reclaimed {
-                // Resume the reclaimed server-side context (full queue intact).
-                let _ = app.svc.engine.play();
-                app.transport.playback_started = true;
-            } else {
-                // No live session — resume the persisted source (context/radio/liked).
-                resume_source(app, &chans.radio);
-                app.transport.playback_started = true;
-            }
-        }
-        KeyCode::Media(MediaKeyCode::Stop) => {
-            app.svc.engine.stop();
-        }
-        KeyCode::Char('n') | KeyCode::Media(MediaKeyCode::TrackNext) => {
-            let _ = app.svc.engine.next();
-        }
-        KeyCode::Char('b') | KeyCode::Media(MediaKeyCode::TrackPrevious) => {
-            let _ = app.svc.engine.prev();
-        }
-        KeyCode::Char('+') | KeyCode::Char('=') | KeyCode::Media(MediaKeyCode::RaiseVolume) => {
-            app.transport.volume = (app.transport.volume + 5).min(100);
-            let _ = app.svc.engine.set_volume(vol_u16(app.transport.volume));
-        }
-        KeyCode::Char('-') | KeyCode::Char('_') | KeyCode::Media(MediaKeyCode::LowerVolume) => {
-            app.transport.volume = app.transport.volume.saturating_sub(5);
-            let _ = app.svc.engine.set_volume(vol_u16(app.transport.volume));
-        }
-        KeyCode::Char('s') => {
-            app.transport.shuffle = !app.transport.shuffle;
-            let _ = app.svc.engine.shuffle(app.transport.shuffle);
-        }
-        // Play the highlighted playlist / album / artist outright. Enter still
-        // opens; this is the direct route that used to require two Enters or
-        // the actions menu.
-        KeyCode::Char('P') => play_selected_context(app, false),
-        KeyCode::Char('S') => {
-            // Flip the global toggle too, or the footer would show shuffle off
-            // while playback is shuffled, and `resume_source` would later
-            // replay this context unshuffled.
-            app.transport.shuffle = true;
-            let _ = app.svc.engine.shuffle(true);
-            play_selected_context(app, true);
-        }
-        KeyCode::Char('R') => {
-            app.transport.repeat = !app.transport.repeat;
-            let _ = app.svc.engine.repeat(app.transport.repeat);
-        }
-        KeyCode::Char('r') => {
-            app.status = "loading library…".to_string();
-            spawn_library_fetch(
-                app.svc.webapi.clone(),
-                chans.lib.clone(),
-                chans.libdone.clone(),
-            );
-        }
-        KeyCode::Char('o') => {
-            app.browse.sort = app.browse.sort.next();
-            let m = app.browse.sort;
-            sort_list(app.cur_list_mut(), m);
-            app.browse.selected = app.first_selectable();
-            app.status = format!("sorted by {}", m.label());
-        }
-        KeyCode::Char('a') => {
-            // Zen hides the library, so the menu belongs to what is playing —
-            // acting on a selection nobody can see is how it ends up offering
-            // "remove from Liked" for the wrong track.
-            let item = if app.view.zen {
-                app.playback
-                    .now
-                    .as_ref()
-                    .filter(|n| !n.uri.is_empty())
-                    .map(|n| LibItem::track(n.title.clone(), n.artist.clone(), n.uri.clone()))
-            } else {
-                app.cur_items().get(app.browse.selected).cloned()
-            };
-            if let Some(item) = item {
-                if !item.is_header && !item.is_play {
-                    // Instant menu (no network), then enrich when the API returns.
-                    app.view.actions = Some(build_action_menu(None, &item));
-                    spawn_action_menu(app.svc.webapi.clone(), item, chans.menu.clone());
-                }
-            }
-        }
-        // Tab / Shift+Tab (and [ ]) rotate the library sections.
-        KeyCode::Tab | KeyCode::Char(']') => {
-            app.search.searching = false;
-            app.browse.section = app.browse.section.shift(1);
-            app.browse.selected = app.first_selectable();
-        }
-        KeyCode::BackTab | KeyCode::Char('[') => {
-            app.search.searching = false;
-            app.browse.section = app.browse.section.shift(-1);
-            app.browse.selected = app.first_selectable();
-        }
-        // Arrow keys rotate the right-pane view; Shift+arrows seek ±5s.
-        KeyCode::Right if mods.contains(KeyModifiers::SHIFT) => {
-            app.playback.seek_step(SEEK_STEP_MS)
-        }
-        KeyCode::Left if mods.contains(KeyModifiers::SHIFT) => {
-            app.playback.seek_step(-SEEK_STEP_MS)
-        }
-        KeyCode::Right => {
-            app.view.mode = app.view.mode.shift(1);
-            if app.view.mode == RightView::Queue
-                && (app.session.reclaimed || app.transport.playback_started)
-            {
-                spawn_queue_fetch(app.svc.webapi.clone(), chans.queue.clone());
-            }
-        }
-        KeyCode::Left => {
-            app.view.mode = app.view.mode.shift(-1);
-            if app.view.mode == RightView::Queue
-                && (app.session.reclaimed || app.transport.playback_started)
-            {
-                spawn_queue_fetch(app.svc.webapi.clone(), chans.queue.clone());
-            }
-        }
-        // The frame loop notices the layout change and wipes the art box.
-        KeyCode::Char('z') => app.view.zen = !app.view.zen,
-        KeyCode::Down | KeyCode::Char('j') => app.move_sel(1),
-        KeyCode::Up | KeyCode::Char('k') => app.move_sel(-1),
-        // Needs a terminal that reports modified Enter (kitty, WezTerm, foot).
-        KeyCode::Enter if mods.contains(KeyModifiers::SHIFT) => play_selected_context(app, false),
-        KeyCode::Enter => match app.activate() {
-            Activated::Open(uri, name) => {
-                spawn_detail_fetch(app.svc.webapi.clone(), uri, name, chans.detail.clone());
-            }
-            Activated::Radio(uri) => {
-                app.status = "starting radio…".to_string();
-                let session = app.svc.engine.session();
-                let tx = chans.radio.clone();
-                tokio::spawn(async move {
-                    let res = match tokio::time::timeout(
-                        Duration::from_secs(12),
-                        engine::radio_tracks(&session, &uri),
-                    )
-                    .await
-                    {
-                        Ok(r) => r.map_err(|e| e.to_string()),
-                        Err(_) => {
-                            Err("timed out (mercury radio endpoint unresponsive)".to_string())
-                        }
-                    };
-                    let _ = tx.send(res.map(|uris| Radio {
-                        uris,
-                        start_position_ms: 0,
-                    }));
-                });
-            }
-            Activated::None => {}
-        },
-        _ => {}
-    }
-    false
-}
-
-/// Keys whose whole effect is on the library pane.
-///
-/// Zen hides that pane, so these do nothing there rather than moving a selection
-/// nobody can see — and `a` is deliberately absent, because it retargets onto
-/// the playing track instead of going quiet.
-fn drives_library(code: KeyCode) -> bool {
-    matches!(
-        code,
-        KeyCode::Tab
-            | KeyCode::BackTab
-            | KeyCode::Up
-            | KeyCode::Down
-            | KeyCode::Enter
-            | KeyCode::Esc
-            | KeyCode::Char('/' | '[' | ']' | 'j' | 'k' | 'o' | 'r' | 'P' | 'S')
-    )
-}
-
-fn consume_media_event<T>(event: Result<T, flume::RecvError>, open: &mut bool) -> Option<T> {
-    match event {
-        Ok(event) => Some(event),
-        Err(_) => {
-            *open = false;
-            None
-        }
-    }
-}
-
-fn handle_media_control_event(
-    app: &mut App,
-    ev: MediaControlEvent,
-    radio_tx: &flume::Sender<Result<Radio, String>>,
-) {
-    match ev {
-        MediaControlEvent::Next => {
-            let _ = app.svc.engine.next();
-        }
-        MediaControlEvent::Previous => {
-            let _ = app.svc.engine.prev();
-        }
-        MediaControlEvent::Toggle => {
-            if app.transport.playback_started {
-                let _ = app.svc.engine.toggle();
-            } else if app.session.reclaimed {
-                // Resume the reclaimed server-side context (full queue intact).
-                let _ = app.svc.engine.play();
-                app.transport.playback_started = true;
-            } else {
-                // No live session — resume the persisted source (context/radio/liked).
-                resume_source(app, radio_tx);
-                app.transport.playback_started = true;
-            }
-        }
-        MediaControlEvent::Play => {
-            if app.transport.playback_started {
-                let _ = app.svc.engine.play();
-            } else if app.session.reclaimed {
-                // Resume the reclaimed server-side context (full queue intact).
-                let _ = app.svc.engine.play();
-                app.transport.playback_started = true;
-            } else {
-                // No live session — resume the persisted source (context/radio/liked).
-                resume_source(app, radio_tx);
-                app.transport.playback_started = true;
-            }
-        }
-        MediaControlEvent::Pause => {
-            let _ = app.svc.engine.pause();
-        }
-        MediaControlEvent::Stop => {
-            app.svc.engine.stop();
-        }
-        MediaControlEvent::Seek(direction) => match direction {
-            SeekDirection::Backward => app.playback.seek_step(-5_000),
-            SeekDirection::Forward => app.playback.seek_step(5_000),
-        },
-        MediaControlEvent::SeekBy(direction, duration) => match direction {
-            SeekDirection::Backward => app.playback.seek_step(-(duration.as_millis() as i64)),
-            SeekDirection::Forward => app.playback.seek_step(duration.as_millis() as i64),
-        },
-        MediaControlEvent::SetPosition(MediaPosition(duration)) => {
-            app.playback
-                .seek_to(&app.svc.engine, duration.as_millis() as u32);
-        }
-        _ => {}
-    }
-}
-
-/// Handle input while the actions menu is open.
-fn handle_action_key(
-    app: &mut App,
-    code: KeyCode,
-    detail_tx: &flume::Sender<(String, String, Vec<LibItem>)>,
-    astatus_tx: &flume::Sender<String>,
-) {
-    match code {
-        KeyCode::Esc | KeyCode::Char('a') => {
-            app.view.actions = None;
-            app.status.clear();
-            return;
-        }
-        KeyCode::Up | KeyCode::Char('k') => {
-            if let Some(m) = app.view.actions.as_mut() {
-                m.selected = m.selected.saturating_sub(1);
-            }
-            return;
-        }
-        KeyCode::Down | KeyCode::Char('j') => {
-            if let Some(m) = app.view.actions.as_mut() {
-                m.selected = (m.selected + 1).min(m.items.len().saturating_sub(1));
-            }
-            return;
-        }
-        KeyCode::Enter => {}
-        _ => return,
-    }
-
-    // Enter: act on the selected entry.
-    let kind = app
-        .view
-        .actions
-        .as_ref()
-        .and_then(|m| m.items.get(m.selected))
-        .map(|i| i.kind.clone());
-    let Some(kind) = kind else { return };
-    match kind {
-        ActionKind::AddToPlaylistMenu { track_uri } => {
-            let items: Vec<ActionItem> = app
-                .browse
-                .library
-                .playlists
-                .iter()
-                .filter_map(|p| {
-                    let id = p.uri.rsplit(':').next()?.to_string();
-                    Some(ActionItem {
-                        label: p.name.clone(),
-                        kind: ActionKind::AddToPlaylist {
-                            playlist_id: id,
-                            track_uri: track_uri.clone(),
-                        },
-                    })
-                })
-                .collect();
-            if items.is_empty() {
-                app.status = "no playlists to add to".to_string();
-                app.view.actions = None;
-            } else {
-                app.view.actions = Some(ActionMenu {
-                    title: "Add to playlist".to_string(),
-                    items,
-                    selected: 0,
-                });
-            }
-        }
-        ActionKind::Play { uri, name } => {
-            // Previously called engine.play_context directly, leaving
-            // source/source_name stale: PLAYING FROM showed the wrong context
-            // and resume-on-launch replayed the previous one.
-            let shuffle = app.transport.shuffle;
-            app.play_context_row(uri, name, shuffle);
-            app.view.actions = None;
-        }
-        ActionKind::Open { uri, name } => {
-            spawn_detail_fetch(app.svc.webapi.clone(), uri, name, detail_tx.clone());
-            app.view.actions = None;
-        }
-        ActionKind::CopyLink { uri } => {
-            app.status = if copy_to_clipboard(&uri_to_url(&uri)) {
-                "link copied".to_string()
-            } else {
-                "clipboard unavailable".to_string()
-            };
-            app.view.actions = None;
-        }
-        other => {
-            spawn_action(app.svc.webapi.clone(), other, astatus_tx.clone());
-            app.view.actions = None;
-        }
-    }
-}
-
-/// Copy text to the system clipboard via whatever tool is available.
-fn copy_to_clipboard(text: &str) -> bool {
-    use std::io::Write;
-    use std::process::{Command, Stdio};
-    let candidates: [(&str, &[&str]); 4] = [
-        ("wl-copy", &[]),
-        ("xclip", &["-selection", "clipboard"]),
-        ("xsel", &["-b", "-i"]),
-        ("pbcopy", &[]),
-    ];
-    for (cmd, args) in candidates {
-        if let Ok(mut child) = Command::new(cmd)
-            .args(args)
-            .stdin(Stdio::piped())
-            .stdout(Stdio::null())
-            .stderr(Stdio::null())
-            .spawn()
-        {
-            if let Some(mut sin) = child.stdin.take() {
-                let _ = sin.write_all(text.as_bytes());
-            }
-            let _ = child.wait();
-            return true;
-        }
-    }
-    false
 }
 
 fn spawn_action_menu(webapi: Arc<Mutex<WebApi>>, item: LibItem, tx: flume::Sender<ActionMenu>) {
